requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
